+++ run.py
@@ -1,37 +1,15 @@
-import click
-from shareFile import create_app, db, bcrypt
+from shareFile import create_app, db, bcrypt, admin
 from shareFile.components.User.model import User
-from shareFile.components.User.auth import is_admin
 from shareFile.components.FileEntry.model import FileEntry, SharedFiles
+from shareFile.components.Admin.view import CustomAdminView
 from config import Config
-from flask_admin import Admin
-from flask import redirect, url_for, request, session
-from flask_admin import AdminIndexView
-from flask_admin.contrib.sqla import ModelView
 
 app = create_app(Config)
-app.config['FLASK_ADMIN_SWATCH'] = 'cyborg'
-
-class MyIndexView(AdminIndexView):
-    def is_accessible(self):
-        return is_admin(session.get("id"))
-    def inaccessible_callback(self, name, **kwargs):
-        return redirect(url_for('user_route.login', next=request.url))
-
-admin = Admin(app, name="shareFile", template_mode="bootstrap3", index_view=MyIndexView())
-
-class CustomAdminView(ModelView):
-    def is_accessible(self):
-        return is_admin(session.get("id"))
-    def inaccessible_callback(self, name, **kwargs):
-        return redirect(url_for('user_route.login', next=request.url))
-
 admin.add_view(CustomAdminView(User, db.session))
 admin.add_view(CustomAdminView(FileEntry, db.session))
 admin.add_view(CustomAdminView(SharedFiles, db.session))
 
 @app.cli.command("initdb")
-#@click.argument("name")
 def initdb():
     db.drop_all()
     db.create_all()
